test: move pipewire recorder coverage
